fix some tests failing during mvn test
