+++ src/test/java/org/prebid/server/it/IntegrationTest.java
@@ -1,5 +1,10 @@
 package org.prebid.server.it;
 
+import com.fasterxml.jackson.databind.JsonNode;
+import com.github.tomakehurst.wiremock.common.FileSource;
+import com.github.tomakehurst.wiremock.extension.Parameters;
+import com.github.tomakehurst.wiremock.extension.ResponseTransformer;
+import com.github.tomakehurst.wiremock.http.Request;
 import com.github.tomakehurst.wiremock.junit.WireMockClassRule;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.config.ObjectMapperConfig;
@@ -7,15 +12,25 @@ import io.restassured.config.RestAssuredConfig;
 import io.restassured.internal.mapping.Jackson2Mapper;
 import io.restassured.response.Response;
 import io.restassured.specification.RequestSpecification;
+import io.vertx.core.json.Json;
+import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.prebid.server.VertxTest;
+import org.prebid.server.cache.proto.request.BidCacheRequest;
+import org.prebid.server.cache.proto.request.PutObject;
+import org.prebid.server.cache.proto.response.BidCacheResponse;
+import org.prebid.server.cache.proto.response.CacheObject;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.TestPropertySource;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
+import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
+import static com.github.tomakehurst.wiremock.client.WireMock.get;
+import static com.github.tomakehurst.wiremock.client.WireMock.urlPathEqualTo;
 import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.options;
 import static java.lang.String.format;
 
@@ -29,7 +44,7 @@ public abstract class IntegrationTest extends VertxTest {
     @SuppressWarnings("unchecked")
     @ClassRule
     public static final WireMockClassRule wireMockRule =
-            new WireMockClassRule(options().port(WIREMOCK_PORT).extensions(ApplicationTest.CacheResponseTransformer.class));
+            new WireMockClassRule(options().port(WIREMOCK_PORT).extensions(IntegrationTest.CacheResponseTransformer.class));
 
     @Rule
     public WireMockClassRule instanceRule = wireMockRule;
@@ -40,6 +55,14 @@ public abstract class IntegrationTest extends VertxTest {
             .setConfig(RestAssuredConfig.config()
                     .objectMapperConfig(new ObjectMapperConfig(new Jackson2Mapper((aClass, s) -> mapper))))
             .build();
+
+    @BeforeClass
+    public static void setUp() throws IOException {
+        wireMockRule.stubFor(get(urlPathEqualTo("/periodic-update"))
+                .willReturn(aResponse().withBody(jsonFrom("storedrequests/test-periodic-refresh.json"))));
+        wireMockRule.stubFor(get(urlPathEqualTo("/currency-rates"))
+                .willReturn(aResponse().withBody(jsonFrom("currency/latest.json"))));
+    }
 
     static String jsonFrom(String file) throws IOException {
         // workaround to clear formatting
@@ -96,5 +119,65 @@ public abstract class IntegrationTest extends VertxTest {
         }
 
         return expectedResponseJson;
+    }
+
+    private static String cacheResponseFromRequestJson(String requestAsString, String requestCacheIdMapFile) throws
+            IOException {
+        List<CacheObject> responseCacheObjects = new ArrayList<>();
+
+        try {
+            final BidCacheRequest cacheRequest = mapper.readValue(requestAsString, BidCacheRequest.class);
+            final JsonNode jsonNodeMatcher =
+                    mapper.readTree(ApplicationTest.class.getResourceAsStream(requestCacheIdMapFile));
+            final List<PutObject> puts = cacheRequest.getPuts();
+
+            for (PutObject putItem : puts) {
+                if (putItem.getType().equals("json")) {
+                    String id = putItem.getValue().get("id").textValue() + "@" + putItem.getValue().get("price");
+                    String uuid = jsonNodeMatcher.get(id).textValue();
+                    responseCacheObjects.add(CacheObject.of(uuid));
+                } else {
+                    String id = putItem.getValue().textValue();
+                    String uuid = jsonNodeMatcher.get(id).textValue();
+                    responseCacheObjects.add(CacheObject.of(uuid));
+                }
+            }
+
+            final BidCacheResponse bidCacheResponse = BidCacheResponse.of(responseCacheObjects);
+            return Json.encode(bidCacheResponse);
+        } catch (IOException e) {
+            throw new IOException("Error while matching cache ids");
+        }
+    }
+
+    public static class CacheResponseTransformer extends ResponseTransformer {
+
+        @Override
+        public com.github.tomakehurst.wiremock.http.Response transform(
+                Request request, com.github.tomakehurst.wiremock.http.Response response, FileSource files,
+                Parameters parameters) {
+
+            final String newResponse;
+            try {
+                newResponse = cacheResponseFromRequestJson(request.getBodyAsString(),
+                        parameters.getString("matcherName"));
+            } catch (IOException e) {
+                return com.github.tomakehurst.wiremock.http.Response.response()
+                        .body(e.getMessage())
+                        .status(500)
+                        .build();
+            }
+            return com.github.tomakehurst.wiremock.http.Response.response().body(newResponse).status(200).build();
+        }
+
+        @Override
+        public String getName() {
+            return "cache-response-transformer";
+        }
+
+        @Override
+        public boolean applyGlobally() {
+            return false;
+        }
     }
 }
